Integrate glossary
Previously, the glossary was available but no files linked to items
in it. Now the glossary has been extended and source files link to
its entries. Note that for the glossary to work, the entries need to
be 3 spaces indented from the left and the text for each entry another
3 spaces.
Note: 2/4 spaces (i.e. 2 for the entries and 4 four the text)
indentation does not work.
* When entries are not indented but the text is, Sphinx errors that no
entries linked in the source files exist in the glossary.
* When entries are indented with 2/4 spaces, the glossary is malformed
and no links can be established.