Split files at chapters to reduce HTML page size
A major redesign of the source structure has been undertaken to optimize (i.e. minimize) the size of individual HTML
pages. In the original setting, only parts had their own folders and index files. Sphinx (and thus ReadTheDocs: RTD)
converts these files into individual HTML pages based on the TOC. For long parts/chapters this becomes an issue as users
have to download the entire part/chapter in their browser, even though they may only be interested in a small section.
It also causes the overall UX to be messy.

To alleviate this issue, the folder structure is now hierarchical, in which each folder has an index.rst file that has
both an introductory text and a hidden TOC that refers to the chapters. These index.rst files are referenced in the main
TOC, that includes all hidden TOCs. We want to show the entries in the main table of contents but not for the individual
chapters.

An additional advantage of this setup is that Google Analytics does not register a bounce whenever a user reads a page
but does not interact with it by clicking on links or (sub-)sections. When the entire HTML page is loaded, users do not
need to interact with the page as it is already available. Now, they have to click on (sub-)sections and/or Next buttons
in RTD to read more. Such navigation is more common and user-friendly anyway, especially on mobile devices.

In addition, several fixes and improvements have been implemented:

* All hyperlink targets are now back-quoted, which ensures that targets with colons do not break.

* Duplicate external hyperlink targets are fixed.

    Whenever Sphinx encounters a hyperlink target with the same reference name, it throws a warning 'duplicate explicit
    target name'. Apart from the fact that these warnings are bogging up the build process screen, they can be avoided
    by using anonymous hyperlinks, which are generally not recommended though: http://tinyurl.com/nbcppyl. The advantage
    of using these is that the warnings go away. The disadvantage is that they must be listed in the order in which they
    appear, which means that placing them in a separate file can become tricky.

    Whenever the hyperlink reference cannot be made more verbose and thus unique, and anonymous hyperlinks must be used,
    these links are placed after the paragraph in which they appear, which takes care of the issue with ordering.

    The issue is also less visible now because of the redesign, as the files themselves are considerably smaller.
    External hyperlinks and their targets must be defined in the same file. Unfortunately, reStructuredText does not
    have a bibliography management solution similar to BibTeX in LaTeX.

* The tocdepth value is larger, which does not affect the RTD sidebar but it is visible in the overview. The sidebar
  only goes two levels deep as specified in the official theme documentation: http://tinyurl.com/kbhbjlj.

* Additional information on cursors is included.

The following command can be run to look for dangling backticks in the final build:
grep -rI '`' build/* | grep -vE '\_sources|\_static|\_images'
These dangling backticks indicate malformed reStructuredText directives (e.g. term and ref) or problems with external
hyperlinks.