Fix issue #8
The HTML translation engine of Sphinx is not 100% compatible with
LaTeX: when curly braces are added to the :math: role these cause
the character to be repeated, most likely due to some HTML/CSS
weirdness that is currently beyond me. Without additional curly
braces the problem goes away in HTML. Note that the issue never
affected the PDFs generated because these are handled completely
by pdflatex.