Split section into separate rst files
To enable easier navigation (and quicker) loading on mobile devices / over bad connections,
it is best to keep the size of each rst file to a minimum. To that end, the section on
deterministic functions and the function result cache has been split into separate files
under the "alternatives" folder.