Fix PL/SQL section hierarchy
The chapter/section hierarchy in the PL/SQL branch of the document was all visible
underneath "Compilation" because the text decorations did not match *and* because
the same folder structure wasn't used. This has been corrected.

Moreover, internal cross-references have been added, especially from the SQL part
to the PL/SQL part. This makes the document more consistent.